implement wrappers to include interactive functionality in the base classes
